backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0

# Validation
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0

# Security (required by PyMySQL)
cryptography==42.0.0
